Moved SampleRelationType schemas and resources to own files. Closes #104
