+++ mass_api_client/schemas/__init__.py
@@ -3,5 +3,6 @@ from .analysis_system import AnalysisSystemSchema
 from .analysis_system_instance import AnalysisSystemInstanceSchema
 from .report import ReportSchema
 from .sample import SampleSchema
-from .sample_relation import SampleRelationSchema, SampleRelationTypeSchema
+from .sample_relation import SampleRelationSchema
+from .sample_relation_type import SampleRelationTypeSchema
 from .scheduled_analysis import ScheduledAnalysisSchema
