+++ mass_api_client/schemas/sample_relation.py
@@ -3,12 +3,6 @@ from marshmallow import fields
 from .base import BaseSchema
 
 
-class SampleRelationTypeSchema(BaseSchema):
-    name = fields.String(required=True)
-    directed = fields.Boolean(required=True)
-    description = fields.String()
-
-
 class SampleRelationSchema(BaseSchema):
     sample = fields.Url(required=True)
     other = fields.Url(required=True)
