+++ mass_api_client/schemas/sample_relation_type.py
@@ -1,0 +1,9 @@
+from marshmallow import fields
+
+from .base import BaseSchema
+
+
+class SampleRelationTypeSchema(BaseSchema):
+    name = fields.String(required=True)
+    directed = fields.Boolean(required=True)
+    description = fields.String()
